test: extract pointer-events tests into pointer-events.test.js
