+++ packages/overlay/test/multiple.test.js
@@ -48,70 +48,6 @@ describe('multiple overlays', () => {
         expect(overlay1._last).to.be.true;
       });
     });
-
-    describe('pointer-events', () => {
-      it('should restore pointer-events correctly when overlays are not closed in order', () => {
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-        expect(document.body.style.pointerEvents).to.eql('none');
-
-        overlay1.opened = false;
-
-        overlay2.opened = false;
-        expect(document.body.style.pointerEvents).to.eql('');
-      });
-
-      it('should disable pointer-events in first overlay when second opens', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('none');
-      });
-
-      it('should restore pointer-events in first overlay when second closes', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        overlay2.opened = false;
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('auto');
-      });
-
-      it('should restore pointer-events in second overlay when third closes', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        const overlay2Part = overlay2.shadowRoot.querySelector('[part="overlay"]');
-        overlay1.opened = true;
-
-        overlay2.opened = true;
-
-        overlay3.opened = true;
-
-        overlay3.opened = false;
-
-        expect(getComputedStyle(overlay2Part).pointerEvents).to.equal('auto');
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('none');
-      });
-
-      it('should clear pointer events after closing overlays', () => {
-        const overlay1Part = overlay1.shadowRoot.querySelector('[part="overlay"]');
-        // Step 1: Opening overlay 1 so it's physically moved under the body
-        overlay1.opened = true;
-        // Step 2: As overlay2 is modal, it will set overlay 1's pointer events to none
-        overlay2.opened = true;
-        // Step 3: Closing overlay 1 so it's physically moved back from under the body
-        overlay1.opened = false;
-        // Step 4: Closing overlay 2 restores pointer-events of an overlay it
-        // finds under the body node, but overlay 1 is no longer there.
-        overlay2.opened = false;
-        // The fix: Clear pointer-events whenever an overlay is closed
-        // (in this case overlay 1 at step 3)
-        expect(getComputedStyle(overlay1Part).pointerEvents).to.equal('auto');
-      });
-    });
   });
 
   describe('modeless', () => {
